@@ -7,10 +7,10 @@
     "@ya.praktikum/react-developer-burger-ui-components": "^1.14.0",
     "clsx": "^1.2.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-intersection-observer": "^8.32.5",
     "react-redux": "^8.1.2",
     "react-router-dom": "^5.3.0",
